mtk: hostapd: update channel information in WBCS to carry maximal BW with no punctured sub-channel.
When AP is about to switch to a new channel with punctured sub-channel,
it will indicated the punctured bitmap in Bandwidth Indication
sub-element of Channel Switch Wrapper. However, the WBCS sub-element has
no field for punctured bitmap and is responsible for carring channel switch
information for legacy STAs, so the AP should calculate the maximal BW
with no punctured sub-channel and carry it in WBCS sub-element.

mtk: hostapd: do not indicate a bandwidth more than 320 MHz in WBCS element

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>