mtk: hostapd: skip BSS from disallowed channel when matching the scan results
In the single wiphy, we have added 'freq_list' in wpa_s configuration to
restrict legacy STA's scan & connection to specified band.
(For example, 2GHz legacy STA only scan 2GHz channels and should only
 connect to a 2GHz AP)

However, the legacy STA migth be configured on each band in a serious of
tests, and the test might also configure the AP with same SSID on each
band. As a result, the STA's scan results contain the BSS with same SSID
but from different bands, which might cause the STA to trigger association
from a different band than the user expected.

This commits skips the BSS from disallowed channel when the STA is
matching the scan results, ensuring that the connection can be triggered
on the correct band.

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>