WNM: Fix pre-scan rejection heuristic for BTM handling
The idea was to only accept the cached scan results if the new target is
reasonably good. To avoid having to write a custom quality logic, a call
to wpa_supplicant_need_to_roam_within_ess() was used. However, the
intention was to swap the parameters and check whether we would want to
roam from the new BSS back to the current one.

Fix the heuristic to match the comment. To do that, we need to add a
parameter to not poll the current signal level as that would result in
comparing the current BSS with itself within the function.

Fixes: 20ed289a785c ("WNM: Clean up old scan data processing")
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>