Fix sae_pk_gen compilation
The Makefile is not properly set up to compile the sae_pk_gen target.
Enabling CONFIG_SAE=y and CONFIG_SAE_PK=y in the configuration is not
sufficient to compile. The linker complains about undefined references.
These additions must be made to the Makefile to get it to compile, at
least with gcc. Once this is done, 'make sae_pk_gen' can be run without
issue.

Signed-off-by: Ameer Antar <ameerantar@comcast.net>