mtk: hostapd: add 'mld_link_id' configuration
Introduce the 'mld_link_id' configuration option to allow users to
specify the link ID for the affiliated AP of the AP MLD.

Please note that the 'mld_link_id' is not validated across links in
hostapd. In other words, users are required to set a unique link ID
for each individual link.

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>