AP: Clean up MLD changes that modified skipping DMG deauthentication
Commit 05e5e615e6a2 ("AP: Skip authentication/deauthentication phase for
DMG/IEEE 802.11ad") added the check for DMG in the beginnign of
ap_sta_deauthenticate() to convert that call to ap_sta_disassociate()
since deauthentication is not used in DMG. Commit c6f519ff15b2 ("AP:
Support deauthenticate/disassociate with MLD") ended up moving this DMG
check into the ap_sta_handle_deauthenticate() function that gets called
once for each link. This is confusing even though DMG is not really used
in MLD.

Move the DMG check back to the beginning of ap_sta_deauthenticate() to
make this clearer.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>