SAE: Avoid duplicated debug entries for IEs in SAE commit messages
Print the "SAE: Possible elements at the end of the frame" debug message
only once (and only if there is actually some additional data) instead
of printing it for each element separately. There was some use for the
separated prints earlier, but that is not really helpful anymore with
the reduced mixing of IEs and non-IE fields at the end of the SAE commit
messages.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>